Extend blog categories that allow the blog categories post only visible to  logged in user

+++ flametreecms/updates/update_blog_plugin_table.php
@@ -3,22 +3,29 @@
 use Schema;
 use October\Rain\Database\Schema\Blueprint;
 use October\Rain\Database\Updates\Migration;
+use System\Classes\PluginManager;
 
 class UpdateBlogPlugin extends Migration
 {
     public function up()
     {
-        Schema::table('rainlab_blog_categories', function (Blueprint $table) {
-            $table->string('featured_image')->nullable()->after('description');
-        });
+        if (PluginManager::instance()->hasPlugin('RainLab.Blog')) {
+            Schema::table('rainlab_blog_categories', function (Blueprint $table) {
+                $table->string('featured_image')->nullable()->after('description');
+                $table->unsignedInteger('user_group')->nullable()->after('id');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('rainlab_blog_categories' , function(Blueprint $table){
-            $table->dropColumn([
-                'featured_image'
-            ]);
-        });
+        if (PluginManager::instance()->hasPlugin('RainLab.Blog')) {
+            Schema::table('rainlab_blog_categories', function (Blueprint $table) {
+                $table->dropColumn([
+                    'featured_image',
+                    'user_group',
+                ]);
+            });
+        }
     }
 }
